Decrease the gas required for the module's integration tests to pass.
* Update funding balances
* Move heavy integration tests to stress tests
